Fixed error in stalemate declaration
Stalemate declaration previously had a syntax error by not including the document object to call createElement(var) in declareStalemate(). The text area has also been changed from a div to a p. Also fixed here is the text in the move list for a stalemate. The text has been fixed to appropriately display a 1/2 symbol.