Fixed layout of outcome declaration text
The text that appeared upon a winner/stalemate being declared was previously to the right of the menu which was unintended. chess.php has added a div, 'outcomediv',  with a p to hold the text for the winner within the overarching 'chesselement' div.
chesslogic.css has added styling for this div to fix the layout. 
chesslogic.js has changed the declareStalemate(), and declareWinner(team,method) functions to not create text nodes but to instead just change the innerHTML of the p within outcomediv.