Fixed error in logic of piece promotion
promotePiece() function was previously missing an == 'white' clause in an if statement, this has been corrected.
recordMove() has had its parameters changed to accommodate for error in logic of piece promotion. It now takes in a flag parameter promotion_flag to see if it was called by promotePiece() or movePiece(), and will use different variables depending on where it was called.